Fix usb_poll type conflict — use unsigned long to match host CC_CHECK
u32 resolves to different types on host gcc -m32 vs cross compiler.
Use explicit unsigned long in both declaration and definition.

+++ lib/netplay/usb.h
@@ -79,7 +79,7 @@
         @return The data header, or 0
     ==============================*/
     
-    extern u32 usb_poll(void);
+    extern unsigned long usb_poll(void);
     
     
     /*==============================
